Update FolderedResourcePackScreen.java
Fixed override setup for available and selected back folders. The button and search field are not wanting to obey resizing conventions, this also includes the remade more packs and done buttons. It looks like Mojang has different layout methods that handle these now. The layout can have items added to it though. I disabled these items for know to get the mod in a presentable state with the folder based back structure working.